Move MCP server to internal/servermcp package, proxy to server proto
Address review feedback:
- Remove InferAuthType middleware; clients must set X-Auth-Type header
- Move MCP server into its own internal/servermcp package with a Server
  struct that takes the XAgentServiceHandler interface
- Proxy tool calls to server proto methods (ListWorkspaces, CreateTask)
  instead of accessing the store directly, reusing existing permissions
- Server struct exposes Handler() returning http.Handler
